seastore: enable cbjournal when running unittest_transaction_manager

+++ src/test/crimson/seastore/transaction_manager_test_state.h
@@ -14,6 +14,8 @@
 #include "crimson/os/seastore/segment_manager.h"
 #include "crimson/os/seastore/collection_manager/flat_collection_manager.h"
 #include "crimson/os/seastore/onode_manager/staged-fltree/fltree_onode_manager.h"
+#include "crimson/os/seastore/random_block_manager/nvmedevice.h"
+#include "crimson/os/seastore/journal/circular_bounded_journal.h"
 
 using namespace crimson;
 using namespace crimson::os;
@@ -23,6 +25,8 @@ class EphemeralTestState {
 protected:
   segment_manager::EphemeralSegmentManagerRef segment_manager;
   std::list<segment_manager::EphemeralSegmentManagerRef> secondary_segment_managers;
+  std::unique_ptr<nvme_device::NVMeBlockDevice> rb_device;
+  journal_type j_type = journal_type::SEGMENT_JOURNAL;
 
   EphemeralTestState(std::size_t num_segment_managers) {
     assert(num_segment_managers > 0);
@@ -58,10 +62,18 @@ protected:
     _mount().handle_error(crimson::ct_error::assert_all{}).get0();
   }
 
-  seastar::future<> tm_setup() {
+  seastar::future<> tm_setup(journal_type j_type = journal_type::SEGMENT_JOURNAL) {
+    this->j_type = j_type;
     segment_manager = segment_manager::create_test_ephemeral();
     for (auto &sec_sm : secondary_segment_managers) {
       sec_sm = segment_manager::create_test_ephemeral();
+    }
+    if (j_type == journal_type::CIRCULARBOUNDED_JOURNAL) {
+      auto config =
+	journal::CircularBoundedJournal::mkfs_config_t::get_default();
+      rb_device.reset(new nvme_device::TestMemory(config.total_size));
+      rb_device->set_device_id(
+	1 << (std::numeric_limits<device_id_t>::digits - 1));
     }
     return segment_manager->init(
     ).safe_then([this] {
@@ -109,6 +121,9 @@ protected:
       for (auto &sec_sm : secondary_segment_managers) {
         sec_sm.reset();
       }
+      if (j_type == journal_type::CIRCULARBOUNDED_JOURNAL) {
+	rb_device.reset();
+      }
     });
   }
 };
@@ -139,8 +154,14 @@ protected:
   TMTestState(std::size_t num_devices) : EphemeralTestState(num_devices) {}
 
   virtual void _init() override {
-    tm = make_transaction_manager(true);
+    tm = make_transaction_manager(true,
+      j_type == journal_type::SEGMENT_JOURNAL ? false : true);
     tm->add_device(segment_manager.get(), true);
+    if (j_type == journal_type::CIRCULARBOUNDED_JOURNAL) {
+      tm->add_device(rb_device.get(), false);
+      static_cast<journal::CircularBoundedJournal*>(tm->get_journal())->
+	add_device(rb_device.get());
+    }
     if (get_num_devices() > 1) {
       for (auto &sec_sm : secondary_segment_managers) {
         tm->add_device(sec_sm.get(), false);
@@ -179,10 +200,24 @@ protected:
   }
 
   virtual FuturizedStore::mkfs_ertr::future<> _mkfs() {
-    return tm->mkfs(
-    ).handle_error(
-      crimson::ct_error::assert_all{"Error in teardown"}
-    );
+    if (j_type == journal_type::SEGMENT_JOURNAL) {
+      return tm->mkfs(
+      ).handle_error(
+	crimson::ct_error::assert_all{"Error in mkfs"}
+      );
+    } else {
+      auto config = journal::CircularBoundedJournal::mkfs_config_t::get_default();
+      return static_cast<journal::CircularBoundedJournal*>(tm->get_journal())->mkfs(
+	config
+      ).safe_then([this]() {
+	return tm->mkfs(
+	).handle_error(
+	  crimson::ct_error::assert_all{"Error in mkfs"}
+	);
+      }).handle_error(
+	crimson::ct_error::assert_all{"Error in mkfs"}
+      );
+    }
   }
 
   auto create_mutate_transaction() {
